Completely separate crud and email packages from gql package
