+++ src/main/java/com/sfjs/email/EmailMessage.java
@@ -1,4 +1,4 @@
-package com.sfjs.gql.schema;
+package com.sfjs.email;
 
 import lombok.Getter;
 import lombok.Setter;
